fixes and comments resolutions, following conventions
